Move torch-native gated delta reference into tests
Review feedback: the pure-torch gated delta implementations are far too
slow for training (serial over documents and sequence steps), so they do
not belong in the model as a runtime backend. Move them into
test_qwen3_5_deltanet.py as the numerical baseline for the FLA kernels,
injected through the GatedDeltaKernel module boundary, and drop
"torch_native" from GatedDeltaBackend. The per-document conv fallback
stays in the model as the non-CUDA path (FLA's causal conv is a triton
kernel).
