+++ torchtitan/models/qwen3_5/model.py
@@ -36,84 +36,7 @@ from .rope import MRoPE
 from .sharding import set_qwen35_sharding_config
 from .vision_encoder import Qwen35VisionEncoder
 
-GatedDeltaBackend = Literal["fla_chunked", "fla_fused_recurrent", "torch_native"]
-
-
-def _l2norm(x: torch.Tensor, dim: int = -1, eps: float = 1e-6) -> torch.Tensor:
-    """L2 norm using rsqrt(sum(x²) + eps), not x/max(norm, eps) like F.normalize, to match FLA kernel."""
-    return x * torch.rsqrt((x * x).sum(dim=dim, keepdim=True) + eps)
-
-
-def _torch_native_gated_delta(
-    q_BLNK: torch.Tensor,
-    k_BLNK: torch.Tensor,
-    v_BLNV: torch.Tensor,
-    g_BLN: torch.Tensor,
-    beta_BLN: torch.Tensor,
-) -> torch.Tensor:
-    """Standalone math reference for the gated delta rule recurrence.
-
-    Sequential O(seqlen) loop — use FLA kernels for GPU efficiency.
-
-    Args:
-        q_BLNK, k_BLNK: (batch, seq, n_heads, key_head_dim)
-        v_BLNV: (batch, seq, n_heads, value_head_dim)
-        g_BLN: (batch, seq, n_heads) -- log-space decay, always negative
-        beta_BLN: (batch, seq, n_heads) -- update gate in (0, 1)
-
-    Returns:
-        output: (batch, seq, n_heads, value_head_dim)
-    """
-    B, L, N, K = q_BLNK.shape
-    V = v_BLNV.shape[-1]
-    dtype = q_BLNK.dtype
-
-    # Upcast to float32 — recurrence accumulates over seqlen steps
-    q_BLNK = _l2norm(q_BLNK.float(), dim=-1) * (K**-0.5)
-    k_BLNK = _l2norm(k_BLNK.float(), dim=-1)
-    v_BLNV, g_BLN, beta_BLN = v_BLNV.float(), g_BLN.float(), beta_BLN.float()
-
-    out_BLNV = torch.zeros(B, L, N, V, dtype=torch.float32, device=q_BLNK.device)
-    state_BNKV = torch.zeros(B, N, K, V, dtype=torch.float32, device=q_BLNK.device)
-
-    for t in range(L):
-        q_BNK = q_BLNK[:, t]
-        k_BNK = k_BLNK[:, t]
-        v_BNV = v_BLNV[:, t]
-        g_BN11 = g_BLN[:, t].exp().unsqueeze(-1).unsqueeze(-1)
-        beta_BN1 = beta_BLN[:, t].unsqueeze(-1)
-
-        state_BNKV = state_BNKV * g_BN11
-        kv_mem_BNV = torch.einsum("bnkv,bnk->bnv", state_BNKV, k_BNK)
-        delta_BNV = (v_BNV - kv_mem_BNV) * beta_BN1
-        state_BNKV = state_BNKV + torch.einsum("bnk,bnv->bnkv", k_BNK, delta_BNV)
-        out_BLNV[:, t] = torch.einsum("bnkv,bnk->bnv", state_BNKV, q_BNK)
-
-    return out_BLNV.to(dtype)
-
-
-def _torch_native_gated_delta_varlen(
-    q_BLNK: torch.Tensor,
-    k_BLNK: torch.Tensor,
-    v_BLNV: torch.Tensor,
-    g_BLN: torch.Tensor,
-    beta_BLN: torch.Tensor,
-    cu_seqlens: torch.Tensor,
-    cu_seqlens_cpu: torch.Tensor,
-) -> torch.Tensor:
-    out_segments_BLNV: list[torch.Tensor] = []
-    cu_seqlens_list = cu_seqlens_cpu.tolist()
-    for start, end in zip(cu_seqlens_list[:-1], cu_seqlens_list[1:], strict=False):
-        out_segments_BLNV.append(
-            _torch_native_gated_delta(
-                q_BLNK[:, start:end],
-                k_BLNK[:, start:end],
-                v_BLNV[:, start:end],
-                g_BLN[:, start:end],
-                beta_BLN[:, start:end],
-            )
-        )
-    return torch.cat(out_segments_BLNV, dim=1)
+GatedDeltaBackend = Literal["fla_chunked", "fla_fused_recurrent"]
 
 
 def _torch_native_causal_conv1d_varlen(
@@ -145,10 +68,11 @@ def _causal_conv1d_varlen(
     weight: torch.Tensor,
     cu_seqlens: torch.Tensor,
     conv_kernel_size: int,
-    backend: GatedDeltaBackend,
     cu_seqlens_cpu: torch.Tensor | None = None,
 ) -> torch.Tensor:
-    if backend == "torch_native" or not x_BTD.is_cuda:
+    # FLA's causal conv is a triton kernel; fall back to per-document
+    # F.conv1d on non-CUDA tensors.
+    if not x_BTD.is_cuda:
         if cu_seqlens_cpu is None:
             raise ValueError(
                 "Qwen3.5 GatedDeltaNet varlen requires CPU cu_seqlens "
@@ -231,19 +155,20 @@ class RMSNormGated(Module):
 
 
 class GatedDeltaKernel(Module):
-    """Stateless dispatch to FLA kernel or pure-torch fallback.
+    """Stateless dispatch to the configured FLA gated delta kernel.
 
     Provides a module boundary for the sharding code to wrap forward with
     DTensor→local conversion — same pattern as FlexAttention. Handles Q/K
     head expansion for grouped linear attention internally so that
-    repeat_interleave runs on local tensors under TP.
+    repeat_interleave runs on local tensors under TP. A pure-torch reference
+    implementation lives in ``tests/unit_tests/test_qwen3_5_deltanet.py``;
+    it is far too slow for training use.
     """
 
     @dataclass(kw_only=True, slots=True)
     class Config(Module.Config):
         # "fla_chunked": parallel within chunks, fast for training (default)
         # "fla_fused_recurrent": token-by-token, lower memory for long sequences
-        # "torch_native": pure-Python reference, for numerical testing only
         backend: GatedDeltaBackend = "fla_chunked"
 
     def __init__(self, config: Config):
@@ -267,30 +192,6 @@ class GatedDeltaKernel(Module):
             repeat = xv_BLNV.shape[2] // xq_BLNK.shape[2]
             xq_BLNK = xq_BLNK.repeat_interleave(repeat, dim=2)
             xk_BLNK = xk_BLNK.repeat_interleave(repeat, dim=2)
-
-        if self.backend == "torch_native":
-            if cu_seqlens is None:
-                return _torch_native_gated_delta(
-                    xq_BLNK,
-                    xk_BLNK,
-                    xv_BLNV,
-                    g_BLN,
-                    beta_BLN,
-                )
-            if cu_seqlens_cpu is None:
-                raise ValueError(
-                    "Qwen3.5 GatedDeltaNet varlen requires CPU cu_seqlens "
-                    "metadata. Build VarlenMetadata with include_host_offsets=True."
-                )
-            return _torch_native_gated_delta_varlen(
-                xq_BLNK,
-                xk_BLNK,
-                xv_BLNV,
-                g_BLN,
-                beta_BLN,
-                cu_seqlens,
-                cu_seqlens_cpu,
-            )
 
         if cu_seqlens is not None and xq_BLNK.shape[0] != 1:
             raise ValueError(
@@ -326,7 +227,7 @@ class GatedDeltaKernel(Module):
         else:
             raise ValueError(
                 f"Unknown fla_backend '{self.backend}'. "
-                "Valid: 'fla_chunked', 'fla_fused_recurrent', 'torch_native'."
+                "Valid: 'fla_chunked', 'fla_fused_recurrent'."
             )
 
         # FLA kernels return (output, final_state); we only need output
@@ -448,7 +349,6 @@ class GatedDeltaNet(Module):
                         w_local,
                         cu_seqlens_local,
                         self.conv_kernel_size,
-                        self.kernel.backend,
                         cu_seqlens_cpu,
                     )
 
@@ -458,7 +358,6 @@ class GatedDeltaNet(Module):
                 conv.weight,
                 cu_seqlens,
                 self.conv_kernel_size,
-                self.kernel.backend,
                 cu_seqlens_cpu,
             )
